Level One layed out
I layed out most of the game objects that need positioned and need art.
Still need to add the puzzles, Back office, and front of building.
Issues, Temporary collider box to keep player within the restraints of the building. probably needs to be replaced with a script as it would be easier than positioning collider boxes everywhere.